Rename dd-foundation to foundation
